+++ frontend/src/App.tsx
@@ -10,8 +10,9 @@ import { Submission } from './model/domain/Submission';
 import { Login } from './components/login/Login';
 import Cookies from 'js-cookie';
 import { ErrorModal } from './components/errorModal/ErrorModal';
-import { Admin } from './components/admin/Admin';
+import { AdminLogin } from './components/admin/AdminLogin';
 import { Stats } from './components/stats/Stats';
+import { Admin } from './components/admin/Admin';
 
 function App() {
   const [semesterOver, setSemesterOver] = useState<boolean>(false);
@@ -69,7 +70,7 @@ function App() {
           setSemesterOver={setSemesterOver}
         />
         <Routes>
-          <Route path="/admin" element={<Admin setErrorMessage={setErrorMessage} />} />
+          <Route path="/admin-login" element={<AdminLogin setErrorMessage={setErrorMessage} />} />
           <Route
             path="/login"
             element={<Login setErrorMessage={setErrorMessage} setUser={setUser} setSubmissions={setSubmissions} setIsAdmin={setIsAdmin} />}
@@ -81,6 +82,7 @@ function App() {
             <Route path="/profile" element={<UserInfo impersonating={impersonating} user={user} setUser={setUser} isAdmin={isAdmin} />} />
             <Route path="/submissions" element={<Submissions submissions={submissions} />} />
             {user.isAdmin && <Route path="/stats" element={<Stats setErrorMessage={setErrorMessage} />} />}
+            {user.isAdmin && <Route path="/admin" element={<Admin />} />}
             <Route path="*" element={<Grader user={user} setSubmissions={setSubmissions} impersonating={impersonating} />} />
           </Routes>
         ) : (
